Fixing squid:S1905 - Redundant casts should not be used

+++ src/main/java/com/fasterxml/jackson/databind/node/FloatNode.java
@@ -73,7 +73,7 @@ public class FloatNode extends NumericNode
     public long longValue() { return (long) _value; }
 
     @Override
-    public float floatValue() { return (float) _value; }
+    public float floatValue() { return _value; }
     
     @Override
     public double doubleValue() { return _value; }
